reorganise vf-gutenberg assets
