+++ wp-content/plugins/vf-gutenberg/blocks/vf-blocks.jsx
@@ -2,9 +2,8 @@
  * VF Gutenberg blocks
  * https://visual-framework.github.io/vf-core/
  */
-import * as button from './blocks/vf-button';
+import * as button from './button';
 
 const {registerBlockType} = wp.blocks;
 
 registerBlockType('vf/button', button.settings);
-
